+++ database/migrations/2019_12_11_040605_crear_dietas_habituales.php
@@ -15,7 +15,7 @@ class CrearDietasHabituales extends Migration
     {
         Schema::create('dietasHabituales', function (Blueprint $table){
             $table->increments('id');
-            $table->smallInteger('consulta_id')->unsigned();
+            $table->integer('consulta_id')->unsigned();
             $table->smallInteger('cereales')->unsigned();
             $table->smallInteger('leguminosas')->unsigned();
             $table->smallInteger('verduras')->unsigned();
